[New Feature] Code refactor DistriOptimizer for advance parameter operations  (#2556)

+++ scala/dllib/src/main/scala/com/intel/analytics/bigdl/dllib/tensor/QuantizedTensorUnsupported.scala
@@ -1442,7 +1442,7 @@ abstract class QuantizedTensorUnsupported[T: ClassTag] extends Tensor[T] {
   override def digamma(): Tensor[T] =
     throw new UnsupportedOperationException(errorString)
 
-  override def clamp(minValue: Float, maxValue: Float): Tensor[T] =
+  override def clamp(minValue: Double, maxValue: Double): Tensor[T] =
     throw new UnsupportedOperationException(errorString)
 
   override def sumSquare(): T =
